zkcc/bip340: extract reusable EC circuit gadgets from Bip340Verify
Add bip340_gadgets.h with a standalone templated Bip340Gadgets class
that exposes the production-tested EC formulas (point-on-curve check,
projective add/double, double-and-add scalar multiplication, scalar
range, bit reconstruction, and even-parity assertion) as public methods.

Bip340Verify now delegates to Bip340Gadgets, ensuring exactly one C++
implementation of each formula is shared by both the native monolithic
circuit and the Lua-authored gadget circuit.
